chore: restructure/refactor
