+++ go.mod
@@ -1,34 +1,25 @@
 module github.com/patoui/logme
 
-go 1.18
+go 1.20
 
 require (
-	github.com/go-chi/chi/v5 v5.0.7
-	github.com/meilisearch/meilisearch-go v0.24.0
-	github.com/stretchr/testify v1.8.2
+	github.com/go-chi/chi/v5 v5.0.8
+	github.com/google/uuid v1.3.0
+	github.com/joho/godotenv v1.5.1
+	github.com/meilisearch/meilisearch-go v0.25.0
+	github.com/mitchellh/mapstructure v1.5.0
+	github.com/stretchr/testify v1.8.4
 )
 
 require (
 	github.com/andybalholm/brotli v1.0.4 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/golang-jwt/jwt/v4 v4.5.0 // indirect
 	github.com/josharian/intern v1.0.0 // indirect
 	github.com/klauspost/compress v1.15.6 // indirect
 	github.com/mailru/easyjson v0.7.7 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/valyala/bytebufferpool v1.0.0 // indirect
 	github.com/valyala/fasthttp v1.37.1-0.20220607072126-8a320890c08d // indirect
-)
-
-require (
-	github.com/ClickHouse/clickhouse-go/v2 v2.0.14
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/google/uuid v1.3.0
-	github.com/joho/godotenv v1.4.0
-	github.com/mitchellh/mapstructure v1.5.0
-	github.com/paulmach/orb v0.7.1 // indirect
-	github.com/pierrec/lz4/v4 v4.1.14 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/shopspring/decimal v1.3.1 // indirect
-	go.opentelemetry.io/otel v1.7.0 // indirect
-	go.opentelemetry.io/otel/trace v1.7.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
