--- internal/logme/home.go
@@ -1,13 +1,0 @@
-package logme
-
-import (
-	"encoding/json"
-	"net/http"
-)
-
-func Home(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(map[string]string{
-		"message": "Welcome to LogMe!",
-	})
-}
