+++ internal/routes/home.go
@@ -1,0 +1,13 @@
+package routes
+
+import (
+    "encoding/json"
+    "net/http"
+)
+
+func Home(w http.ResponseWriter, r *http.Request) {
+    w.Header().Set("Content-Type", "application/json")
+    json.NewEncoder(w).Encode(map[string]string{
+        "message": "Welcome to LogMe!",
+    })
+}
